Refactor InputStyler, move array logic to ArrayFormatter component.
